requirements.txt
python-telegram-bot
plexapi
pyyaml
